// Projects/Tamin/Eservices-UI/src/app/models/insurance-list/listDetail.model.ts
export class ListDetailModel  {
    workshopId: string;
    listNumber: string;
    workshopTitle: string;
    workshopCode: string;
    yearAndMonthDescription: string;
    totalEmployeePremium: string;
    totalEmployerPremium: string;
    hardJobPremium: string;
    totalUnemploymentPremium: string;
    employeesCount: string;
    totalInclusiveBenefit: string;
    exclusive: string;
    totalWageAndBenefit: string;
    totalWorkingDays: string;
    totalDailyWages: string;
    totalMonthlyWages: string;
    listDetailModel: string;
    totalInclusiveWageAndBenefit: string;
    contractNumber: string;
    id: string;
    employerName: string;
    listDescription: string;
    listKind: string;
    pourcentage: string;
    premiumRate: string;
    status: string;
    workshopAddress: string;
    brchCode: string;
    year: string;
    month: {
        code:string
        monthDescription:string
    };

     }